Refactor bot into package
