+++ bot/queue.py
@@ -1,0 +1,7 @@
+from collections import deque
+
+number_queue = deque()
+user_queue = deque()
+bindings = {}
+blocked_numbers = {}
+IGNORED_TOPICS = set()
